Use SupportedPlugin.OnCall detecting OnCall types as a single source of truth
