+++ public/app/features/alerting/unified/types/pluginBridges.ts
@@ -1,0 +1,5 @@
+export enum SupportedPlugin {
+  Incident = 'grafana-incident-app',
+  OnCall = 'grafana-oncall-app',
+  MachineLearning = 'grafana-ml-app',
+}
